fix RecordsContainer functions
