+++ src/components/RecordPicker.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react'
-import { Link, withRouter } from 'react-router-dom'
+import { withRouter } from 'react-router-dom'
 import { exportCSV } from '../functions/csv'
 import download from '../functions/download'
 import {
@@ -13,6 +13,7 @@ import Pager from './Pager'
 
 import ButtonContiner from './ButtonContainer'
 import SearchRecords from './SearchRecords'
+import RecordsContainer from './RecordsContainer'
 
 /**
  * Renders a list of the available records.
@@ -119,30 +120,6 @@ function RecordPicker(props) {
 
 	function onExactMatchChanged(e) {
 		setState({ ...state, exactMatch: e.target.checked, page: 0 })
-	}
-
-	function getList(str) {
-		return str
-			.match(/(".*?"|[^",\s]+)(?=\s*,|\s*$)/g)
-			.map((d) => d.replace(/"/g, ''))
-	}
-
-	function getFieldText(codebook, record, fieldName) {
-		let field = codebook.find((d) => d.name === fieldName)
-		if (
-			field.type === 'qualitative' &&
-			field.valid_values &&
-			field.value_labels
-		) {
-			let values = getList(field.valid_values).map((d) => parseInt(d))
-			let labels = getList(field.value_labels)
-			let value = parseInt(record[fieldName])
-			return labels[values.find((d) => d.toString() === value.toString())]
-		}
-
-		if (isUnknown(record[fieldName] || '', field)) return 'unknown'
-
-		return record[fieldName]
 	}
 
 	function onPageChange(page) {
@@ -259,66 +236,7 @@ function RecordPicker(props) {
 	let pageCount = Math.ceil(filteredRecords.length / state.pageSize)
 	let page = Math.max(0, Math.min(state.page, pageCount - 1))
 
-	let records = filteredRecords
-		// Paging
-		.slice(state.page * state.pageSize, (state.page + 1) * state.pageSize)
-		.map((d) => {
-			let validation = validateRecord(d, props.codebook)
-			let issues = Object.keys(validation)
-				.filter((d) => !validation[d].valid)
-				.map((d) => [d, validation[d]])
-
-			let issueDisplay = null
-
-			if (issues.length > 0) {
-				issueDisplay = (
-					<span className="issues">
-						<span className="fa fa-warning issues" />
-						<span>{issues.length}</span>
-					</span>
-				)
-			}
-
-			var locked = (d.locked || '').toString().toLowerCase() === 'true'
-
-			return (
-				<Link
-					key={d.uid}
-					to={'/record/' + d.uid}
-					className={`list-item ${locked ? ' locked' : ''}`}
-				>
-					<span className="pid">
-						{locked && <span className="fa fa-lock"> </span>} {d.pid}{' '}
-						{issueDisplay}
-					</span>
-					<span className="hits">
-						{searchHits[d.uid] &&
-							searchHits[d.uid].slice(0, 10).map((e, i) => (
-								<span key={i}>
-									{e[0]}: {e[1]}
-								</span>
-							))}
-					</span>
-					<span className="sort-field">
-						{getFieldText(props.codebook, d, state.sortField)}
-					</span>
-					<button
-						disabled={d.locked === 'TRUE'}
-						onClick={(e) => {
-							e.preventDefault()
-							deleteRecord(d.uid)
-						}}
-						className={`button ${
-							d.locked === 'TRUE' ? 'is-disabled' : ' is-danger'
-						} is-small is-outlined is-rounded remove`}
-					>
-						<span className="fa fa-remove" />
-					</button>
-				</Link>
-			)
-		})
-
-	var sort = (
+	const sort = (
 		<div className="sort">
 			<Pager
 				page={page}
@@ -405,7 +323,17 @@ function RecordPicker(props) {
 				searchField={state.searchField}
 			/>
 			{sort}
-			<div className="list is-hoverable">{records}</div>
+			<div className="list is-hoverable">
+				<RecordsContainer
+					page={state.page}
+					pageSize={state.pageSize}
+					codebook={props.codebook}
+					filteredRecords={filteredRecords}
+					searchHits={searchHits}
+					deleteRecord={deleteRecord}
+					sortField={state.sortField}
+				/>
+			</div>
 		</div>
 	)
 }
